# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Postos API",
        "description": "API de gestão de postos (PinPoint): JWT, RBAC, stock e histórico.\r\n\r\n**Rotas públicas (sem JWT):**\r\n- `GET /postos` — listar, paginar (`?page=`) ou detalhe (`?id=`)\r\n- `GET /postos/search` — pesquisa rápida (`?q=`, opcional `limit`, `tipo`, `combustivel`, `status`)\r\n\r\n**Papéis locais (`users.role_id`):** ver `docs/ROLES.md` e `GET /api/roles`.",
        "contact": {
            "email": "suporte@postosapi.local"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "Prefixo das rotas API"
        }
    ],
    "paths": {
...
        "/postos/search": {
            "get": {
                "tags": [
                    "Postos Públicos"
                ],
                "summary": "Pesquisar postos (público, rápido)",
                "description": "**Rota pública** — não requer `Authorization`.\n\nPesquisa em `nome`, `endereco` e `combustivel` (apenas postos activos).\n\n- `q`: obrigatório, mínimo 2 caracteres\n- `limit`: opcional (default 20, máximo 50)\n- Filtros opcionais: `tipo`, `combustivel`, `status`\n- Resposta em cache ~60 segundos\n\nExemplo: `GET /api/postos/search?q=shell&status=aberto&limit=10`",
                "operationId": "8fb713bcd00ed77ef0ad5ef91996b491",
                "parameters": [
                    {
                        "name": "q",
                        "in": "query",
                        "description": "Termo de pesquisa (nome, morada ou combustível)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "shell",
                            "minLength": 2
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Máximo de resultados",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 20,
                            "default": 20,
                            "maximum": 50
                        }
                    },
                    {
                        "name": "tipo",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "combustivel"
                        }
                    },
                    {
                        "name": "combustivel",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "Gasolina"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "aberto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de postos que correspondem à pesquisa",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PostoSearchResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validação falhou (ex.: q em falta ou &lt; 2 caracteres)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The q field is required."
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/postos": {
            "get": {
                "tags": [
                    "Postos Públicos"
                ],
                "summary": "Listar postos (público)",
                "description": "**Rota pública** — não requer `Authorization`.\n\n- Sem parâmetros: todos os postos activos (array)\n- `?id=1`: um posto (`PostoPublic`)\n- `?page=2`: paginação, 10 por página (`PostoPublicPaginatedResponse`)\n\nPara pesquisa textual use `GET /postos/search?q=...`.",
                "operationId": "0d8c4c8337fe50c98d64db4223fe70c5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "ID do posto",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Número da página (10 postos por página)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1,
                            "minimum": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Posto único, lista ou página paginada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "$ref": "#/components/schemas/PostoPublic"
                                        },
                                        {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/PostoPublic"
                                            }
                                        },
                                        {
                                            "$ref": "#/components/schemas/PostoPublicPaginatedResponse"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Posto não encontrado"
                    }
                }
            }
        },
...
        "/posts/{post}/products/{product}/stock": {
            "patch": {
                "tags": [
                    "Stock"
                ],
                "summary": "Atualizar stock de um produto no posto (gestor)",
                "description": "O gestor só pode atualizar stock do posto ao qual está vinculado (user.post_id === post.id).",
                "operationId": "4d17443126e49193e050ad8b066af699",
                "parameters": [
                    {
                        "name": "post",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "product",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateStockRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Stock atualizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/StockUpdateResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado"
                    },
                    "403": {
                        "description": "Sem permissão",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validação falhou"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "PostoSearchResponse": {
                "required": [
                    "q",
                    "count",
                    "postos"
                ],
                "properties": {
                    "q": {
                        "description": "Termo pesquisado",
                        "type": "string",
                        "example": "shell"
                    },
                    "count": {
                        "description": "Número de resultados devolvidos",
                        "type": "integer",
                        "example": 2
                    },
                    "postos": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/PostoPublic"
                        }
                    }
                },
                "type": "object"
            },
            "PostoPublicPaginatedResponse": {
                "properties": {
                    "postos": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/PostoPublic"
                        }
                    },
                    "page": {
                        "type": "integer",
                        "example": 1
                    },
                    "per_page": {
                        "type": "integer",
                        "example": 10
                    },
                    "total": {
                        "type": "integer",
                        "example": 25
                    },
                    "last_page": {
                        "type": "integer",
                        "example": 3
                    }
                },
                "type": "object"
            },
...
            "PostoPublic": {
                "properties": {
                    "id": {
                        "type": "string",
                        "example": "1"
                    },
                    "nome": {
                        "type": "string",
                        "example": "Posto Shell"
                    },
                    "endereco": {
                        "type": "string",
                        "example": "Av. 4 de Fevereiro, 123"
                    },
                    "tipo": {
                        "type": "string",
                        "example": "combustivel"
                    },
                    "coordinates": {
                        "properties": {
                            "latitude": {
                                "type": "number",
                                "format": "float",
                                "example": -8.8383
                            },
                            "longitude": {
                                "type": "number",
                                "format": "float",
                                "example": 13.2344
                            }
                        },
                        "type": "object"
                    },
                    "preco": {
                        "type": "string",
                        "example": "320 Kz/L",
                        "nullable": true
                    },
                    "precoPremium": {
                        "type": "string",
                        "example": "340 Kz/L",
                        "nullable": true
                    },
                    "combustivel": {
                        "type": "string",
                        "example": "Gasolina",
                        "nullable": true
                    },
                    "combustiveis": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "tipo": {
                                    "type": "string",
                                    "example": "gasolina"
                                },
                                "nome": {
                                    "type": "string",
                                    "example": "Gasolina"
                                },
                                "disponibilidade": {
                                    "type": "string",
                                    "example": "em_stock"
                                }
                            },
                            "type": "object"
                        }
                    },
                    "status": {
                        "type": "string",
                        "example": "aberto"
                    },
                    "hours24": {
                        "type": "boolean",
                        "example": true
                    },
                    "image": {
                        "type": "string",
                        "nullable": true
                    },
                    "services": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "example": [
                            "Wifi",
                            "Lavagem",
                            "Loja de Conveniência"
                        ]
                    }
                },
                "type": "object"
            },
...
            "Stock": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 10
                    },
                    "post_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "product_id": {
                        "type": "integer",
                        "example": 42
                    },
                    "quantity": {
                        "type": "number",
                        "example": 150
                    },
                    "critical_level": {
                        "type": "number",
                        "example": 20,
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "UpdateStockRequest": {
                "required": [
                    "quantity"
                ],
                "properties": {
                    "quantity": {
                        "type": "number",
                        "example": 150,
                        "minimum": 0
                    },
                    "critical_level": {
                        "type": "number",
                        "example": 20,
                        "nullable": true,
                        "minimum": 0
                    },
                    "justificativa_ajuste": {
                        "type": "string",
                        "nullable": true,
                        "maxLength": 1000
                    }
                },
                "type": "object"
            },
            "StockUpdateResponse": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Stock updated"
                    },
                    "stock": {
                        "$ref": "#/components/schemas/Stock"
                    }
                },
                "type": "object"
            },
            "ErrorResponse": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Forbidden"
                    },
                    "code": {
                        "type": "integer",
                        "example": 403
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Token JWT no header Authorization: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}